Check if workflow can be restarted for both MoveTables and Workflow start
